+++ projects/core/dxt-lossless-transform-bc3/benches/transform_standard/avx2.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc3::transforms::standard::transform::bench::u32_avx2_transform;
+use dxt_lossless_transform_bc3::transform::standard::transform::bench::u32_avx2_transform;
 use safe_allocator_api::RawAlloc;
 
 fn bench_avx2(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
