+++ projects/core/dxt-lossless-transform-bc3/benches/untransform_standard/sse2.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc3::transforms::standard::untransform::bench::{
+use dxt_lossless_transform_bc3::transform::standard::untransform::bench::{
     u32_untransform_sse2, u64_untransform_sse2,
 };
 use safe_allocator_api::RawAlloc;
