+++ projects/core/dxt-lossless-transform-bc3/src/transform/standard/transform/avx2.rs
