+++ projects/core/dxt-lossless-transform-bc3/src/transform/standard/transform/avx512.rs
