+++ projects/core/dxt-lossless-transform-bc3/src/transform/standard/transform/portable32.rs
