+++ projects/core/dxt-lossless-transform-bc3/src/transform/standard/untransform/avx512.rs
@@ -6,7 +6,7 @@ use core::arch::x86::*;
 use core::arch::x86_64::*;
 use std::is_x86_feature_detected;
 
-use crate::transforms::standard::untransform::portable::u32_untransform_with_separate_pointers;
+use crate::transform::standard::untransform::portable::u32_untransform_with_separate_pointers;
 
 /// # Safety
 ///
